add classes, moved in dedicated folder extensions, fix typos
add custom errors

delete static function to throw errors

moved new files in darwin and add symbolic link

relative path

drop underscore in resume player func
