+++ just_audio/darwin/Classes/EffectData.swift
@@ -1,0 +1,12 @@
+//
+//  EffectData.swift
+//  just_audio
+//
+//  Created by kuama on 22/08/22.
+//
+
+import Foundation
+
+protocol EffectData {
+    var type: EffectType { get }
+}
