+++ just_audio/darwin/Classes/EffectType.swift
@@ -1,0 +1,12 @@
+//
+//  EffectType.swift
+//  just_audio
+//
+//  Created by kuama on 22/08/22.
+//
+
+import Foundation
+
+enum EffectType: String, Codable {
+    case darwinEqualizer = "DarwinEqualizer"
+}
