+++ just_audio/darwin/Classes/Errors/NotImplementedError.swift
@@ -1,0 +1,14 @@
+//
+//  NotImplementedError.swift
+//  just_audio
+//
+//  Created by kuama on 22/08/22.
+//
+
+import Foundation
+
+class NotImplementedError: PluginError {
+    init(_ message: String) {
+        super.init(500, message)
+    }
+}
