+++ just_audio/darwin/Classes/IndexedAudioSource.swift
@@ -2,7 +2,7 @@ import AVFoundation
 
 class IndexedAudioSource: AudioSource {
     func load(engine _: AVAudioEngine, playerNode _: AVAudioPlayerNode, speedControl _: AVAudioUnitVarispeed, position _: CMTime?, completionHandler _: @escaping () -> Void) throws {
-        throw PluginError.notImplemented("Not implemented IndexedAudioSource.load")
+        throw NotImplementedError("Not implemented IndexedAudioSource.load")
     }
 
     func getDuration() -> CMTime {
