+++ just_audio/darwin/Classes/LoopMode.swift
@@ -1,0 +1,12 @@
+//
+//  LoopMode.swift
+//  just_audio
+//
+//  Created by kuama on 22/08/22.
+//
+
+import Foundation
+
+enum LoopMode: Int {
+    case loopOff, loopOne, loopAll
+}
