+++ just_audio/darwin/Classes/ParamsEqualizerData.swift
@@ -1,0 +1,12 @@
+//
+//  ParamsEqualizerData.swift
+//  just_audio
+//
+//  Created by kuama on 22/08/22.
+//
+
+import Foundation
+
+struct ParamsEqualizerData: Codable {
+    let bands: [BandEqualizerData]
+}
