--- just_audio/ios/Classes/JustAudioPlayer.swift
@@ -1,639 +1,0 @@
-import AVFoundation
-
-class PluginError: Error {
-    let code: Int
-    let message: String
-
-    init(_ code: Int, _ message: String) {
-        self.code = code
-        self.message = message
-    }
-
-    static func notImplemented(_ message: String) -> PluginError {
-        return PluginError(500, message)
-    }
-
-    static func notInitialized(_ message: String) -> PluginError {
-        return PluginError(403, message)
-    }
-
-    static func notSupported(_ value: Any, _ message: Any) -> PluginError {
-        return PluginError(400, "Not support \(value)\n\(message)")
-    }
-}
-
-public class JustAudioPlayer: NSObject {
-    let playerId: String
-    let audioEffects: [[String: Any]]
-
-    let methodChannel: FlutterMethodChannel
-    let eventChannel: BetterEventChannel
-    let dataChannel: BetterEventChannel
-
-    var player: Player!
-
-    init(registrar: FlutterPluginRegistrar, playerId: String, loadConfiguration _: [String: Any], audioEffects: [[String: Any]]) {
-        self.playerId = playerId
-        self.audioEffects = audioEffects
-
-        methodChannel = FlutterMethodChannel(name: String(format: "com.ryanheise.just_audio.methods.%@", playerId), binaryMessenger: registrar.messenger())
-        eventChannel = BetterEventChannel(name: String(format: "com.ryanheise.just_audio.events.%@", playerId), messenger: registrar.messenger())
-        dataChannel = BetterEventChannel(name: String(format: "com.ryanheise.just_audio.data.%@", playerId), messenger: registrar.messenger())
-
-        super.init()
-        methodChannel.setMethodCallHandler { call, result in
-            self.handleMethodCall(call: call, result: result)
-        }
-    }
-
-    func handleMethodCall(call: FlutterMethodCall, result: @escaping FlutterResult) {
-        do {
-            if player == nil {
-                player = Player(audioEffects: try! audioEffects.map(Mapping.effectFrom), onEvent: onPlaybackEvent)
-            }
-
-            let request = call.arguments as! [String: Any]
-            print("=========== \(call.method) \(request)")
-
-            switch call.method {
-            case "load":
-                let source = try AudioSource.fromJson(request["audioSource"] as! [String: Any])
-                let initialPosition = request["initialPosition"] != nil ? CMTime.invalid : CMTimeMake(value: request["initialPosition"] as! Int64, timescale: 1_000_000)
-                let initialIndex = request["initialIndex"] as? Int ?? 0
-
-                let duration = player.load(source: source, initialPosition: initialPosition, initialIndex: initialIndex)
-                result(["duration": duration.microSeconds])
-            case "play":
-                player.play()
-                result([:])
-            case "pause":
-                player.pause()
-                result([:])
-            case "stop":
-                player.stop()
-                result([:])
-            case "setVolume":
-                player.setVolume(Float(request["volume"] as! Double))
-                result([:])
-            case "setPitch":
-                player.setPitch(Float(request["pitch"] as! Double))
-                result([:])
-            case "setSkipSilence":
-                // TODO: player.setSkipSilence(request["enabled"] as! Bool)
-                result(PluginError.notImplemented(call.method))
-            case "setSpeed":
-                player.setSpeed(Float(request["speed"] as! Double))
-                result([:])
-            case "setLoopMode":
-                player.setLoopMode(mode: Mapping.loopModeFrom(request["loopMode"] as! Int))
-                result([:])
-            case "setShuffleMode":
-                // random or normal
-                player.setShuffleMode(isEnalbed: Mapping.shuffleModeFrom(request["shuffleMode"] as! Int))
-                result([:])
-            case "setShuffleOrder":
-                // TODO: TEST
-                player.setShuffleOrder(data: request["audioSource"] as! [String: Any])
-                result([:])
-            case "setAutomaticallyWaitsToMinimizeStalling":
-                // android is still to be implemented too
-                result(PluginError.notImplemented(call.method))
-            case "setCanUseNetworkResourcesForLiveStreamingWhilePaused":
-                // even android is still to be implemented too
-                result(PluginError.notImplemented(call.method))
-            case "setPreferredPeakBitRate":
-                // even android is still to be implemented too
-                result(PluginError.notImplemented(call.method))
-            case "setClip":
-                // even android is still to be implemented too
-                result(PluginError.notImplemented(call.method))
-            case "seek":
-                let position = Mapping.timeFrom(microseconds: request["position"] as! Int64)
-                let index = request["index"] as? Int
-
-                player.seek(index: index, position: position)
-                result([:])
-            case "concatenatingInsertAll":
-                result(PluginError.notImplemented(call.method))
-            case "concatenatingRemoveRange":
-                result(PluginError.notImplemented(call.method))
-            case "concatenatingMove":
-                result(PluginError.notImplemented(call.method))
-            case "audioEffectSetEnabled":
-                try player.enableEffect(type: request["type"] as! String, enabled: request["enabled"] as! Bool)
-                result([:])
-            case "darwinEqualizerBandSetGain":
-                player.setEqualizerBandGain(bandIndex: request["bandIndex"] as! Int, gain: Float(request["gain"] as! Double))
-                result([:])
-            default:
-                result(FlutterMethodNotImplemented)
-            }
-        } catch let error as PluginError {
-            result(FlutterError(code: "\(error.code)", message: error.message, details: nil))
-        } catch {
-            print(error)
-            result(FlutterError(code: "500", message: error.localizedDescription, details: nil))
-        }
-    }
-
-    func onPlaybackEvent(event: PlaybackEvent) {
-        eventChannel.sendEvent([
-            "processingState": event.processingState.rawValue,
-            "updatePosition": event.updatePosition.microSeconds,
-            "updateTime": event.updateTime,
-            "bufferedPosition": 0,
-            "icyMetadata": [:],
-            "duration": event.duration.microSeconds,
-            "currentIndex": event.currentIndex,
-        ])
-    }
-
-    func dispose() {
-        player?.dispose()
-        player = nil
-
-        eventChannel.dispose()
-        dataChannel.dispose()
-        methodChannel.setMethodCallHandler(nil)
-    }
-}
-
-enum ProcessingState: Int, Codable {
-    case none, loading, buffering, ready, completed
-}
-
-enum LoopMode: Int {
-    case loopOff, loopOne, loopAll
-}
-
-class Player {
-    let onEvent: (PlaybackEvent) -> Void
-    let audioEffects: [EffectData]
-
-    var engine: AVAudioEngine!
-    var playerNode: AVAudioPlayerNode!
-    var speedControl: AVAudioUnitVarispeed!
-    var pitchControl: AVAudioUnitTimePitch!
-    var audioUnitEQ: AVAudioUnitEQ?
-
-    // State properties
-    var processingState: ProcessingState = .none
-    var shuffleModeEnabled = false
-    var loopMode: LoopMode = .loopOff
-
-    // Queue properties
-    var indexedAudioSources: [IndexedAudioSource] = []
-    var currentSource: IndexedAudioSource?
-    var order: [Int] = []
-    var orderInv: [Int] = []
-
-    // Current Source
-    var index: Int = 0
-    var audioSource: AudioSource!
-    var duration: CMTime {
-        if processingState == .none || processingState == .loading {
-            return CMTime.invalid
-        } else if indexedAudioSources.count > 0 {
-            return currentSource!.getDuration()
-        } else {
-            return CMTime.zero
-        }
-    }
-
-    // Positions properties
-    var positionUpdatedAt: Int64 = 0
-    var positionUpdate: CMTime = .zero
-    var positionOffset: CMTime = .zero
-    var currentPosition: CMTime { return positionUpdate + positionOffset }
-
-    // Extra properties
-    var volume: Float = 1
-    var pitch: Float = 1
-    var rate: Float = 1
-
-    init(audioEffects: [EffectData], onEvent: @escaping (PlaybackEvent) -> Void) {
-        self.audioEffects = audioEffects
-        self.onEvent = onEvent
-    }
-
-    func load(source: AudioSource, initialPosition _: CMTime, initialIndex: Int) -> CMTime {
-        if playerNode != nil {
-            playerNode.pause()
-        }
-
-        index = initialIndex
-        processingState = .loading
-        updatePosition(CMTime.zero)
-        // Decode audio source
-        audioSource = source
-
-        indexedAudioSources = audioSource.buildSequence()
-
-        updateOrder()
-
-        if indexedAudioSources.isEmpty {
-            processingState = .none
-            broadcastPlaybackEvent()
-
-            return CMTime.zero
-        }
-
-        if engine == nil {
-            engine = AVAudioEngine()
-            playerNode = AVAudioPlayerNode()
-            speedControl = AVAudioUnitVarispeed()
-            pitchControl = AVAudioUnitTimePitch()
-
-            try! createAudioEffects()
-
-            playerNode.volume = volume
-            speedControl.rate = rate
-            pitchControl.pitch = pitch
-
-            var nodes = [playerNode, speedControl, pitchControl]
-
-            // add equalizer node
-            if audioUnitEQ != nil {
-                nodes.append(audioUnitEQ!)
-            }
-
-            // attach all nodes to engine
-            for node in nodes {
-                engine.attach(node!)
-            }
-
-            // add mainMixerNode
-            nodes.append(engine.mainMixerNode)
-
-            for i in 1 ..< nodes.count {
-                engine.connect(nodes[i - 1]!, to: nodes[i]!, format: nil)
-            }
-
-            // Observe for changes in the audio engine configuration
-            NotificationCenter.default.addObserver(self,
-                                                   selector: #selector(_handleInterruption),
-                                                   name: NSNotification.Name.AVAudioEngineConfigurationChange,
-                                                   object: nil)
-        }
-
-        try! setQueueFrom(index)
-
-        loadCurrentSource()
-
-        if !engine.isRunning {
-            try! engine.start()
-        }
-
-        processingState = .ready
-        broadcastPlaybackEvent()
-
-        return duration
-    }
-
-    @objc func _handleInterruption(notification _: Notification) {
-        _resume()
-    }
-
-    func play() {
-        playPlayerNode()
-        updatePosition(nil)
-        broadcastPlaybackEvent()
-    }
-
-    func pause() {
-        updatePosition(nil)
-        playerNode.pause()
-        broadcastPlaybackEvent()
-    }
-
-    func stop() {
-        stopPlayerNode()
-        updatePosition(nil)
-        broadcastPlaybackEvent()
-    }
-
-    func _resume() {
-        let wasPlaying = playerNode.isPlaying
-
-        playerNode.pause()
-        if !engine.isRunning {
-            try! engine.start()
-        }
-
-        if wasPlaying {
-            playerNode.play()
-        }
-    }
-
-    func seek(index: Int?, position: CMTime) {
-        let wasPlaying = playerNode.isPlaying
-
-        if let index = index {
-            try! setQueueFrom(index)
-        }
-
-        stopPlayerNode()
-
-        updatePosition(position)
-
-        processingState = .ready
-
-        loadCurrentSource()
-
-        // Restart play if player was playing
-        if wasPlaying {
-            playPlayerNode()
-        }
-
-        broadcastPlaybackEvent()
-    }
-
-    func updatePosition(_ positionUpdate: CMTime?) {
-        positionUpdatedAt = Int64(Date().timeIntervalSince1970 * 1000)
-        if let positionUpdate = positionUpdate { self.positionUpdate = positionUpdate }
-        positionOffset = indexedAudioSources.count > 0 && positionUpdate == nil ? playerNode.currentTime : CMTime.zero
-    }
-
-    private var isStopping = false
-    // Permit to check if [load(completionHandler)] is called when you force a stop
-    private func stopPlayerNode() {
-        isStopping = true
-        playerNode.stop()
-        isStopping = false
-    }
-
-    private func playPlayerNode() {
-        if !engine.isRunning {
-            try! engine.start()
-        }
-        playerNode.play()
-    }
-
-    private func loadCurrentSource() {
-        try! currentSource!.load(engine: engine, playerNode: playerNode, speedControl: speedControl, position: positionUpdate, completionHandler: {
-            if self.isStopping { return }
-            DispatchQueue.main.async {
-                self.playNext()
-            }
-        })
-    }
-
-    private func getRelativeIndex(_ offset: Int) -> Int {
-        switch loopMode {
-        case .loopOne:
-            return index
-        case .loopAll:
-            return offset >= indexedAudioSources.count ? 0 : orderInv[offset]
-        case .loopOff:
-            return orderInv[offset]
-        }
-    }
-
-    private func playNext() {
-        let newIndex = index + 1
-        if newIndex >= indexedAudioSources.count {
-            complete()
-        } else {
-            seek(index: getRelativeIndex(newIndex), position: CMTime.zero)
-            play()
-        }
-    }
-
-    private func complete() {
-        updatePosition(nil)
-        processingState = .completed
-        if playerNode != nil {
-            playerNode.stop()
-        }
-        broadcastPlaybackEvent()
-    }
-
-    // ========== QUEUE
-
-    fileprivate func setQueueFrom(_ index: Int) throws {
-        guard !indexedAudioSources.isEmpty else {
-            preconditionFailure("no songs on library")
-        }
-        self.index = index
-        currentSource = indexedAudioSources[index]
-    }
-
-    // ========== MODES
-
-    fileprivate func setShuffleMode(isEnalbed: Bool) {
-        shuffleModeEnabled = isEnalbed
-        updateOrder()
-        broadcastPlaybackEvent()
-    }
-
-    fileprivate func setShuffleOrder(data: [String: Any]) {
-        audioSource = try! .fromJson(data)
-        switch data["type"] as! String {
-        case "concatenating":
-            let children = (data["children"] as! [[String: Any]])
-            for child in children {
-                setShuffleOrder(data: child)
-            }
-        case "looping":
-            setShuffleOrder(data: data["child"] as! [String: Any])
-        default:
-            break
-        }
-
-        print(audioSource.sourceId)
-    }
-
-    fileprivate func setLoopMode(mode: LoopMode) {
-        loopMode = mode
-        broadcastPlaybackEvent()
-    }
-
-    fileprivate func updateOrder() {
-        orderInv = Array(repeating: 0, count: indexedAudioSources.count)
-        if shuffleModeEnabled {
-            order = audioSource.getShuffleIndices()
-        } else {
-            order = indexedAudioSources.enumerated().map { index, _ in
-                index
-            }
-        }
-        for i in 0 ..< indexedAudioSources.count {
-            orderInv[order[i]] = i
-        }
-    }
-
-    // ========== EFFECTS
-
-    fileprivate func createAudioEffects() throws {
-        for effect in audioEffects {
-            if let effect = effect as? EqualizerEffectData {
-                audioUnitEQ = AVAudioUnitEQ(numberOfBands: effect.parameters.bands.count)
-
-                for (i, band) in effect.parameters.bands.enumerated() {
-                    audioUnitEQ!.bands[i].filterType = .parametric
-                    audioUnitEQ!.bands[i].frequency = band.centerFrequency
-                    audioUnitEQ!.bands[i].bandwidth = 1 // half an octave
-                    audioUnitEQ!.bands[i].gain = Mapping.gainFrom(band.gain)
-                    audioUnitEQ!.bands[i].bypass = false
-                }
-
-                audioUnitEQ!.bypass = !effect.enabled
-            } else {
-                throw PluginError.notSupported(effect.type, "When initialize effect")
-            }
-        }
-    }
-
-    fileprivate func enableEffect(type: String, enabled: Bool) throws {
-        switch type {
-        case "DarwinEqualizer":
-            audioUnitEQ!.bypass = !enabled
-        default:
-            throw PluginError.notInitialized("Not initialized effect \(type)")
-        }
-    }
-
-    fileprivate func setEqualizerBandGain(bandIndex: Int, gain: Float) {
-        audioUnitEQ?.bands[bandIndex].gain = gain
-    }
-
-    // ======== EXTRA
-
-    fileprivate func setVolume(_ value: Float) {
-        volume = value
-        if playerNode != nil {
-            playerNode.volume = volume
-        }
-        broadcastPlaybackEvent()
-    }
-
-    fileprivate func setPitch(_ value: Float) {
-        pitch = value
-        if pitchControl != nil {
-            pitchControl.pitch = pitch
-        }
-        broadcastPlaybackEvent()
-    }
-
-    fileprivate func setSpeed(_ value: Float) {
-        rate = value
-        if speedControl != nil {
-            speedControl.rate = rate
-        }
-        updatePosition(nil)
-    }
-
-    fileprivate func broadcastPlaybackEvent() {
-        onEvent(PlaybackEvent(
-            processingState: processingState,
-            updatePosition: currentPosition,
-            updateTime: positionUpdatedAt,
-            duration: duration,
-            currentIndex: index
-        ))
-    }
-
-    fileprivate func dispose() {
-        if processingState != .none {
-            playerNode?.pause()
-            processingState = .none
-        }
-        audioSource = nil
-        indexedAudioSources = []
-        playerNode?.stop()
-        engine?.stop()
-    }
-}
-
-extension CMTime {
-    var milliSeconds: Int64 {
-        return self == CMTime.invalid ? -1 : Int64(value * 1000 / Int64(timescale))
-    }
-
-    var microSeconds: Int64 {
-        return self == CMTime.invalid ? -1 : Int64(value * 1_000_000 / Int64(timescale))
-    }
-}
-
-extension AVAudioPlayerNode {
-    var currentTime: CMTime {
-        if let nodeTime: AVAudioTime = lastRenderTime, let playerTime: AVAudioTime = playerTime(forNodeTime: nodeTime) {
-            let currentTime = Double(playerTime.sampleTime) / playerTime.sampleRate
-            let milliSeconds = Int64(currentTime * 1000)
-            return milliSeconds < 0 ? CMTime.zero : CMTime(value: milliSeconds, timescale: 1000)
-        }
-        return CMTime.zero
-    }
-}
-
-enum Mapping {
-    static func timeFrom(microseconds: Int64) -> CMTime {
-        return CMTimeMake(value: microseconds, timescale: 1_000_000)
-    }
-
-    static func loopModeFrom(_ value: Int) -> LoopMode {
-        switch value {
-        case 1:
-            return LoopMode.loopOne
-        case 2:
-            return LoopMode.loopAll
-        default:
-            return LoopMode.loopOff
-        }
-    }
-
-    static func shuffleModeFrom(_ value: Int) -> Bool {
-        return value == 1
-    }
-
-    static func gainFrom(_ value: Float) -> Float {
-        // Equalize the level between iOS and android
-        return value * 2.8
-    }
-
-    static func effectFrom(_ map: [String: Any]) throws -> EffectData {
-        let type = map["type"] as! String
-        switch type {
-        case EffectType.darwinEqualizer.rawValue:
-            return EqualizerEffectData.fromJson(map)
-        default:
-            throw PluginError.notSupported(type, "When decoding effect")
-        }
-    }
-}
-
-enum EffectType: String, Codable {
-    case darwinEqualizer = "DarwinEqualizer"
-}
-
-protocol EffectData {
-    var type: EffectType { get }
-}
-
-struct EqualizerEffectData: EffectData, Codable {
-    let type: EffectType
-    let enabled: Bool
-    let parameters: ParamsEqualizerData
-
-    static func fromJson(_ map: [String: Any]) -> EqualizerEffectData {
-        return try! JSONDecoder().decode(EqualizerEffectData.self, from: JSONSerialization.data(withJSONObject: map))
-    }
-}
-
-struct ParamsEqualizerData: Codable {
-    let bands: [BandEqualizerData]
-}
-
-struct BandEqualizerData: Codable {
-    let index: Int
-    let centerFrequency: Float
-    let gain: Float
-}
-
-struct PlaybackEvent {
-    let processingState: ProcessingState
-    let updatePosition: CMTime
-    let updateTime: Int64
-    let duration: CMTime
-    let currentIndex: Int
-}
