--- just_audio/ios/Classes/SwiftJustAudioPlugin.swift
@@ -1,45 +1,0 @@
-import Flutter
-import UIKit
-
-public class SwiftJustAudioPlugin: NSObject, FlutterPlugin {
-    var players: [String: JustAudioPlayer] = [:]
-    let registrar: FlutterPluginRegistrar
-
-    init(registrar: FlutterPluginRegistrar) {
-        self.registrar = registrar
-    }
-
-    public static func register(with registrar: FlutterPluginRegistrar) {
-        let channel = FlutterMethodChannel(name: "com.ryanheise.just_audio.methods", binaryMessenger: registrar.messenger())
-        let instance = SwiftJustAudioPlugin(registrar: registrar)
-        registrar.addMethodCallDelegate(instance, channel: channel)
-    }
-
-    public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
-        switch call.method {
-        case "init":
-            let request = call.arguments as! [String: Any]
-            let playerId = request["id"] as! String
-
-            print("init: ", request)
-
-            let loadConfiguration = request["audioLoadConfiguration"] as? [String: Any] ?? [:]
-            let audioEffects = request["darwinAudioEffects"] as? [[String: Any]] ?? []
-            if players[playerId] != nil {
-                let flutterError = FlutterError(code: "error", message: "Platform player already exists", details: nil)
-                result(flutterError)
-            } else {
-                let player = JustAudioPlayer(registrar: registrar, playerId: playerId, loadConfiguration: loadConfiguration, audioEffects: audioEffects)
-                players[playerId] = player
-                result(nil)
-            }
-        case "disposePlayer":
-            let request = call.arguments as! [String: Any]
-            let playerId = request["id"] as! String
-            players.removeValue(forKey: playerId)?.dispose()
-            result([:])
-        default:
-            result(FlutterMethodNotImplemented)
-        }
-    }
-}
