+++ just_audio/macos/Classes
@@ -1,0 +1,1 @@
+../darwin/Classes
