--- just_audio/macos/Classes/JustAudioPlugin.h
@@ -1,5 +1,0 @@
-#import <Flutter/Flutter.h>
-#import <AVFoundation/AVFoundation.h>
-
-@interface JustAudioPlugin : NSObject<FlutterPlugin>
-@end
